Introduce an upgrader based on NuGet packages
Implement an upgrade mechanism that makes the upgrade logic more flexible to
enable customized steps to be run during an upgrade. Organizations might have
specific requirements and need custom steps to deploy VFSForGit upgrades.

This introduces the NuGet upgrader, which allows for generalized upgrade
packages. The upgrade package is distributed via NuGet feeds as a NuGet
package. The package includes details on how to install itself, allowing
organizations the ability to control the VFSForGit upgrade.