ProductUpgrader: always query GitHub releases for updates
The original design of the NuGetUpgrader was to query the local credential
store for credentials to use to query for updates. This approach had problems
with retreiving credentials, and exposed cases where it would be easier to leak
credentials.

Instead, we are always going to query the GitHub Release feed for updates, even
if the update is downloaded from a NuGet feed.