Upgrader: platform specific blocking process list
Different platforms have different requirements for the list of processes that
block installation.

The POSIX platforms also remove "gvfs" from the list of blocked processes, as
they can run upgrade while the 'gvfs upgrade' command is running. We might want
to come back and restrict this to only the `gvfs upgrade` command (or only if
the running command is a parent process) in a future change.