Upgrader: add command install action
When running install actions, we previously only supported starting an
application or document contained in the installer package. You might
want to run instead a command that is provided by the host operating
system / environment. For example, on macOS you might want to run bash
or the "installer" binary directly.